sql: cleanup handling of EXPLAIN in the new factory
Previously, several EXPLAIN variants were mishandled in the new factory:
instead of being wrapped as other planNodes are they were kept separate
which resulted in a "mixed"-representation plan. This is now fixed by
properly wrapping them.

Additionally, this commit removes "partial support" that we had for
EXPLAIN (PLAN) - the new factory cannot currently implement it, so in
such cases we will be falling back to the old factory.

Release note: None
